pyproject.toml
[project]
name = "frank"
version = "0.1.0"
description = "Frank chat"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "logfire[fastapi]>=4.0.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.4.9",
    "python-dotenv>=1.1.1",
    "upstash-redis>=1.4.0",
]

[dependency-groups]
dev = [
    "clai>=0.4.9",
    "pydantic-to-typescript>=2.0.0",
    "modal>=1.1.0",
]
